Improved the Functionality of isPrime Function.
I improved the functionality of the isPrime Function to return Boolean values instead of just printing them so that it can also be used on other cases (i.e. if-else statements and while loops). This also compliments to printing True or False values. Also, I changed the capitalization of the isPrime Function to Camel Case just to cope up with the common convention🙂.